feat(synth): polyphonic engine (6-voice pool)
Replace the mono single-voice globals with a Voice[6] pool: note-on
allocates (free, else steal oldest), note-off gates the matching
voice. Per-voice phase/env/velocity; global pitch-bend + vibrato LFO;
the low-pass filter and a soft limiter move to the summed master bus
so chords don't clip while one voice stays at the v0.1.0 level.

Also fix oscillator.h: std::sinf → std::sin (std::sinf is non-standard;
Arduino toolchain lacks it; host g++ accepted it, embedded build did not).

Assisted by AI.

+++ src/synth/oscillator.h
@@ -34,7 +34,7 @@ inline float shapeGain(WaveShape shape) {
 
 inline float osc(WaveShape shape, float phase) {
   switch (shape) {
-    case WaveShape::Sine:   return std::sinf(phase);
+    case WaveShape::Sine:   return std::sin(phase);
     case WaveShape::Saw:
       // Rises linearly from -1 at phase=0 to +1 at phase approaching 2π.
       return 2.0f * phase / TWO_PI_F - 1.0f;
